src/test_gdrive.py
import unittest
from unittest.mock import MagicMock, patch
from google.colab import auth, drive
from googleapiclient.discovery import build
from googleapiclient.http import MediaIoBaseDownload
from io import BytesIO
import logging

from gdrive import authenticate_google_drive, mount_google_drive, download_file_from_drive

class TestAuthenticateGoogleDrive(unittest.TestCase):

    @patch('google.colab.auth.authenticate_user')
    @patch('builtins.logging.error')
    def test_authenticate_google_drive_successful(self, mock_log_error, mock_auth):
        # Arrange
        mock_auth.return_value = None

        # Act
        authenticate_google_drive()

        # Assert
        mock_log_error.assert_not_called()

    @patch('google.colab.auth.authenticate_user')
    @patch('builtins.logging.error')
    def test_authenticate_google_drive_error(self, mock_log_error, mock_auth):
        # Arrange
        mock_auth.side_effect = Exception('Authentication error')

        # Act & Assert
        with self.assertRaises(Exception):
            authenticate_google_drive()
        mock_log_error.assert_called_once_with("Error authenticating to Google Drive: Authentication error")

    # Add more test cases for different scenarios

class TestMountGoogleDrive(unittest.TestCase):

    @patch('google.colab.drive.mount')
    @patch('builtins.logging')
    def test_mount_google_drive_successful(self, mock_logging, mock_mount):
        # Arrange
        mock_mount.return_value = None

        # Act
        mount_google_drive()

        # Assert
        mock_logging.info.assert_called_once_with("Successfully mounted Google Drive to /content/drive")

    # Add more test cases for different scenarios

class TestDownloadFileFromDrive(unittest.TestCase):

    @patch('googleapiclient.discovery.build')
    def test_download_file_from_drive_successful(self, mock_build):
        # Arrange
        mock_service = MagicMock()
        mock_build.return_value = mock_service
        file_id = '12345'

        mock_downloader = MagicMock()
        mock_downloader.next_chunk.return_value = (None, True)

        # Act
        result = download_file_from_drive(mock_service, file_id)

        # Assert
        self.assertIsInstance(result, BytesIO)

    # Add more test cases for different scenarios

if __name__ == '__main__':
    unittest.main()
